requirements.txt
fastapi[all]==0.110.0
uvicorn==0.27.1
pytest==7.4.4
pytest-cov==4.1.0
psycopg2-binary==2.9.9
logtail-python